+++ rust/kona/sp1/crates/ethereum/client/src/lib.rs
@@ -3,3 +3,6 @@
 pub mod executor;
 pub mod super_consolidation;
 pub mod super_range;
+
+#[cfg(test)]
+mod test_utils;
